==> 01_get_title_links.ipynb <==
import os
import string
import glob
import pandas as pd 
from bs4 import BeautifulSoup
from csv import writer

path = "lexile/raw/lists/"
files = [f for f in glob.glob(path + "*.html", recursive=True)]
base_url = 'https://www.commonsensemedia.org'

titles = []
links = []
urls = []
file_names = []

def get_titles(soup):
    for s in soup.findAll(class_="views-field views-field-field-reference-review-ent-prod result-title"):
        titles.append(s.get_text().strip())
    return titles

def get_href(soup):
    for s in soup.findAll(class_="views-field views-field-field-reference-review-ent-prod result-title"):
        links.append(s.a)
    return links

def make_urls(links):
    for l in links: 
        l = str(l)
        urls.append(base_url + l.split('"',2)[1])
    return urls

def make_f_names(links):
    for l in links:
        l = str(l)
        l = l.split('/')[2]
        file_names.append(l.split('"')[0])
    return file_names


for file in files:
    with open(file, 'r') as f: 
        soup = BeautifulSoup(f.read(), 'html.parser')
        get_titles(soup)
        get_href(soup)

urls = make_urls(links)
f_names = make_f_names(links)

df = pd.DataFrame()
df['title'] = titles
df['f_name'] = f_names
df['url'] = urls
df.head()

## Save to csv

!mkdir csv

df.to_csv("~/final_project/csv/title-url.csv", index=False)
