==> 02_get_book_reviews.ipynb <==
import pandas as pd
import csv
from bs4 import BeautifulSoup
import requests
import os

df = pd.read_csv('csv/title-url.csv')
df.head()

df.shape[0]

!mkdir lexile/raw/book-reviews

path = 'lexile/raw/book-reviews/'

for i, j in df.iterrows():
    if not os.path.isfile(path + j.f_name + ".html"):
        try: 
            page = requests.get(j.url)
            soup = BeautifulSoup(page.text, 'html.parser')
            with open(path + j.f_name + ".html", 'wb') as f:
                try:
                    f.write(soup.encode('utf-8'))
                except:
                    with open("lexile/raw/book-errors.txt", 'a+') as f:
                        f.write(j.title + "\n")
        except: 
            with open("lexile/raw/book-errors.txt", 'a+') as f:
                        f.write(j.title + "\n")
    else: 
        continue

len(os.listdir(path))
