==> 03_get_book_details.ipynb <==
import pandas as pd
import numpy as np
import csv
from bs4 import BeautifulSoup
import os

file = "lexile/raw/book-reviews/3-willows.html"
with open(file, 'r') as f: 
        soup = BeautifulSoup(f.read(), 'html.parser')

## Get All Data

title = soup.h1.text
description = soup.find(class_='field field-name-field-one-liner field-type-text field-label-hidden').get_text()
csm_age = soup.find(class_="csm-green-age").get_text().replace("age ", "")[:-1]

print("Title: {}".format(title))
print()
print("Description: {}".format(description))
print()
print("CSM Recommended Age: {}".format(csm_age))

### Plot Summary

def get_summary(soup):
    remove_charachters = ["\xa0\n", "\xa0a", "\xa0"]
    plot_sum = soup.find(class_="field field-name-field-what-is-story field-type-text-long field-label-hidden").get_text()
    for charachter in remove_charachters:
        plot_sum = plot_sum.replace(charachter, " ")
    return plot_sum

### Is It Any Good? 

def get_csm_review(soup):
    remove_charachters = ["\xa0\n", "\xa0a", "\xa0"]
    csm_evaluation = soup.find(class_="field field-name-field-any-good field-type-text-long field-label-hidden").get_text()[:-1]
    for charachter in remove_charachters:
        csm_evaluation = csm_evaluation.replace(charachter, " ")
    return csm_evaluation

### Parent's rating

def parents_rating(soup): 
    try:
        for s in soup.find_all(class_='user-review-statistics adult'):
            adult_rating = s.find("div", attrs={'class': 'stat-wrapper age'})
        return adult_rating.text.replace("age ", "")[:-1]
    except:
        return np.nan
parents_rating(soup)

### Children's rating

def kids_rating(soup):
    try:
        for s in soup.find_all(class_='user-review-statistics child'):
            childs_rating = s.find("div", attrs={'class': 'stat-wrapper age'})
        return childs_rating.text.replace("age ", "")[:-1]
    except:
        return np.nan
kids_rating(soup)

def pntk(soup):
    remove_charachters = ["\xa0\n", "\xa0a", "\xa0"]
    pntk = soup.find(class_="field field-name-field-parents-need-to-know field-type-text-long field-label-hidden").get_text()
    for charachter in remove_charachters:
        pntk = pntk.replace(charachter, " ")
    return pntk
pntk(soup)

path = 'lexile/raw/book-reviews/'
files = os.listdir(path)
files = [path + file for file in files]

titles = []
descriptions = [] 
plot_sums = []
csm_evaluations = []
need_to_know = []
p_rating = []
k_rating = []
csm_ages = []

for file in files:
    try: 
        with open(file, 'r') as f: 
            soup = BeautifulSoup(f.read(), 'html.parser')
            titles.append(soup.h1.text)
            descriptions.append(soup.find(class_='field field-name-field-one-liner field-type-text field-label-hidden').get_text())
            csm_ages.append(soup.find(class_="csm-green-age").get_text().replace("age ", "")[:-1])
            plot_sums.append(get_summary(soup))
            csm_evaluations.append(get_csm_review(soup))
            need_to_know.append(pntk(soup))
            p_rating.append(parents_rating(soup))
            k_rating.append(kids_rating(soup))
    except:
        continue

df = pd.DataFrame({
    'title': titles,
    'description' : descriptions,
    'plot': plot_sums,
    'csm_review' : csm_evaluations,
    'need_to_know' : need_to_know,
    'par_rating' : p_rating,
    'kids_rating' : k_rating,
    'csm_rating' : csm_ages,
})

df.tail()

df.shape

df.to_csv('csv/book_info.csv')

## Get All Data from the Summary box

data = []

for file in files:
    with open(file, 'r') as f: 
        soup = BeautifulSoup(f.read(), 'html.parser') 
        summary = soup.find(class_="shutter-summary-pane panel-pane pane-product-details")
        links = summary.find_all('li')
        d = {}
        for link in links:
            k, v = link.text.split(":")
            k = k.strip()
            v = v.strip()
            d[k] = v
        data.append(d)


df2 = pd.DataFrame(data)

df2.head()

df2.shape

df2.to_csv('csv/book_info_summary.csv')

df2.tail()

df3 = pd.concat([df, df2], axis=1)

df.shape

df2.shape

df3.shape

df3.head()

df3.to_csv('csv/book_info_complete.csv', index=False)
