==> 07_xgboost.ipynb <==
import pandas as pd
import numpy as np
import xgboost as xgb

from sklearn.model_selection import StratifiedShuffleSplit
from sklearn.metrics import mean_absolute_error
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.preprocessing import OneHotEncoder 
from sklearn.preprocessing import MinMaxScaler


import nltk
#nltk.download('stopwords')
from nltk.corpus import stopwords
stop = stopwords.words('english')

# [Tutorial](https://machinelearningmastery.com/gradient-boosting-with-scikit-learn-xgboost-lightgbm-and-catboost/)

## Load the Data 

#Run Local
df = pd.read_csv("csv/book_info_complete.csv")
#Run on COLAB
#df = pd.read_csv("/content/drive/My Drive/final_project/book_info_complete.csv")

df = df.dropna(axis=1)
df["Publication date"] = df["Publication date"].str[-4:].astype(int)
df = df.drop("Last updated", axis=1)
df.info()

### Lower and remove stop words

columns = ['title', "description", "plot", "csm_review", "need_to_know"]

df = df.applymap(lambda x: x.lower() if type(x) == str else x)

df[columns] = df[columns].applymap(lambda x: ' '.join([item for item in x.split() if item not in stop]))

df.head(1)

### Make the split

def splitter(df):
    split = StratifiedShuffleSplit(n_splits=1, test_size=0.2, random_state=999)
    for train_index, test_index in split.split(df, df['csm_rating']):
        train_data= df.loc[train_index]
        test_data = df.loc[test_index]
    
    
    return train_data, test_data

train_data, test_data = splitter(df)

train_data.shape

test_data.shape

## One Hot Encode

### Book Type

train_data["Book type"].value_counts()

book_type_cat = train_data[["Book type"]]
book_type_encoder = OneHotEncoder(handle_unknown="ignore")
book_type_cat_1hot = book_type_encoder.fit_transform(book_type_cat)

train_book_type_df = pd.DataFrame(data=book_type_cat_1hot.todense(), 
                            index= train_data.index,
                            columns=book_type_encoder.get_feature_names())

test_book_type_df = pd.DataFrame(data = book_type_encoder.transform(test_data[["Book type"]]).todense(), 
                                  index=test_data.index,
                                  columns=book_type_encoder.get_feature_names())

## Genre

train_data["Genre"].hist(xrot=90);

book_genre_cat = train_data[["Genre"]]
book_genre_encoder = OneHotEncoder(handle_unknown='ignore')
book_genre_cat_1hot = book_genre_encoder.fit_transform(book_genre_cat)

train_book_genre_df = pd.DataFrame(book_genre_cat_1hot.todense(), 
                             index=train_data.index,
                             columns=book_genre_encoder.get_feature_names())

test_book_genre_df = pd.DataFrame(data = book_genre_encoder.transform(test_data[["Genre"]]).todense(), 
                                  index=test_data.index,
                                  columns=book_genre_encoder.get_feature_names())

## MinMaxScale Publication Date

book_pub_year = train_data[["Publication date"]]
min_max_scaler = MinMaxScaler().fit(book_pub_year)
bp_year_MM = min_max_scaler.transform(book_pub_year)

train_bp_year_df = pd.DataFrame(data=bp_year_MM,
                          index=train_data.index, 
                          columns=["book_pub_year"])

test_bp_year_df = pd.DataFrame(data=min_max_scaler.transform(test_data[["Publication date"]]),
                              index=test_data.index,
                              columns=["book_pub_year"])

target = train_data[["csm_rating"]]
min_max_scaler_target = MinMaxScaler().fit(target)
target_MM = min_max_scaler_target.transform(target)

train_target_df = pd.DataFrame(data=target_MM,
                              index=train_data.index,
                              columns=["csm_rating_scaled"])

test_target_df = pd.DataFrame(data=min_max_scaler_target.transform(test_data[["csm_rating"]]),
                              index=test_data.index,
                              columns=["csm_rating_scaled"])



## Create BOW

def generate_bow(x):
    """Create a bag of words and return the vecotorizer along with data frames"""
    #instantiate CountVectorizer()
    cv=CountVectorizer(min_df=2)
    # this step generates word counts for the words in your docs
    word_count_vector=cv.fit_transform(train_data[str(x)])
    
    train = pd.DataFrame(word_count_vector.todense(), 
                      index=train_data.index, 
                      columns=cv.get_feature_names())
    
    test = pd.DataFrame(cv.transform(test_data[str(x)]).todense(), 
                       index=test_data.index,
                       columns=cv.get_feature_names())

    return train, test

train_description_df, test_description_df  = generate_bow('description')
train_title_df, test_title_df = generate_bow("title")
train_plot_df, test_plot_df = generate_bow("plot")
train_csm_review_df, test_csm_review_df = generate_bow("csm_review")
train_need_to_know_df, test_need_to_know_df = generate_bow("need_to_know")

print(train_description_df.index.equals(train_need_to_know_df.index))
print(test_description_df.index.equals(test_title_df.index))

## Create Create Train and Test Data Frames

train_x = pd.concat(objs=[train_description_df, train_bp_year_df, train_book_genre_df,
                         train_book_type_df],
                    axis=1)

train_y = train_data['csm_rating']

test_x = pd.concat(objs=[test_description_df, test_bp_year_df, test_book_genre_df,
                        test_book_type_df],
                    axis=1)

test_y = test_data['csm_rating']

xg_reg = xgb.XGBRegressor(objective ='reg:squarederror', 
                          colsample_bytree = 0.9046665663213616,
                          learning_rate = 0.23377134043030423,
                          gamma=0.9737555188414592,
                          max_depth = 62, 
                          alpha = 1, 
                          n_estimators = 224, 
                          subsample = 0.9618386009333088,
                          booster = "gblinear")

#Parameters: {'colsample_bytree': 0.9046665663213616, 'gamma': 0.9737555188414592,
#             'learning_rate': 0.23377134043030423, 'max_depth': 62, 'n_estimators': 224, 'subsample': 0.9618386009333088}


#Fit the model with the training data
xg_reg.fit(train_x, train_y)

#Predict the taget on the train data set
predict_train = xg_reg.predict(train_x)
predict_test = xg_reg.predict(test_x)

mean_absolute_error(train_y, predict_train)

mean_absolute_error(test_y, predict_test)

### Transform and Predict on the Test DF

#Predict the target on the test dataset
predict_test = xg_reg.predict(test_x)

mean_absolute_error(predict_test, test_y)

test_errors = abs(predict_test - test_y)

predictions  = pd.Series(predict_test, index=test_data.index, name="predictions")

test_errors.name = "difference"

test_results = pd.concat([test_data, test_errors, predictions], axis=1)

test_results.head()

test_results.predictions.unique()
